Move AutoMapperConfiguration to Nop.Core
