+++ src/Libraries/Nop.Core/packages.config
@@ -2,6 +2,7 @@
 <packages>
   <package id="Autofac" version="4.1.0" targetFramework="net451" />
   <package id="Autofac.Mvc5" version="4.0.0" targetFramework="net451" />
+  <package id="AutoMapper" version="5.1.1" targetFramework="net451" />
   <package id="Microsoft.AspNet.Mvc" version="5.2.3" targetFramework="net451" />
   <package id="Microsoft.AspNet.Razor" version="3.2.3" targetFramework="net451" />
   <package id="Microsoft.AspNet.WebPages" version="3.2.3" targetFramework="net451" />
